batch 상용 적용

+++ domain-user/src/main/java/co/dalicious/domain/user/util/PointUtil.java
@@ -4,13 +4,14 @@ import co.dalicious.domain.user.dto.PointPolicyResDto;
 import co.dalicious.domain.user.entity.PointHistory;
 import co.dalicious.domain.user.entity.PointPolicy;
 import co.dalicious.domain.user.entity.User;
+import co.dalicious.domain.user.entity.enums.PointStatus;
 import co.dalicious.domain.user.entity.enums.ReviewPointPolicy;
 import co.dalicious.domain.user.mapper.PointMapper;
 import co.dalicious.domain.user.repository.PointHistoryRepository;
 import co.dalicious.domain.user.repository.QPointHistoryRepository;
-import co.dalicious.domain.user.repository.QPointPolicyRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
@@ -21,13 +22,12 @@ import java.util.stream.Collectors;
 @Component
 @RequiredArgsConstructor
 public class PointUtil {
-
-    private final QPointPolicyRepository qPointPolicyRepository;
     private final QPointHistoryRepository qPointHistoryRepository;
     private final PointHistoryRepository pointHistoryRepository;
     private final PointMapper pointMapper;
 
-    public BigDecimal findReviewPoint(Boolean isReviewImage, BigDecimal itemPrice) {
+    // 리뷰 작성 시 포인트 산정
+    public BigDecimal findReviewPoint(Boolean isReviewImage, BigDecimal itemPrice, int count) {
         BigDecimal reviewPoint = BigDecimal.ZERO;
 
         // 어떤 조건에 부합하는지 찾기
@@ -41,63 +41,69 @@ public class PointUtil {
             Integer max = reviewPointPolicy.getMaxPrice() == null ? null : reviewPointPolicy.getMaxPrice();
             int price = itemPrice.intValue();
 
-            if(min == null && max != null && price < max) {
+            if(min == null && max != null && price <= max) {
                 rewardPointForImage = rewardPointForImage.add(reviewPointPolicy.getImagePoint());
                 rewardPointForContent = rewardPointForContent.add(reviewPointPolicy.getContentPoint());
             }
-            else if(min != null && max == null && min < price) {
+            else if(min != null && max == null && min <= price) {
                 rewardPointForImage = rewardPointForImage.add(reviewPointPolicy.getImagePoint());
                 rewardPointForContent = rewardPointForContent.add(reviewPointPolicy.getContentPoint());
             }
-            else if(min != null && max != null && min < price && max > price) {
+            else if(min != null && max != null && min <= price && max >= price){
                 rewardPointForImage = rewardPointForImage.add(reviewPointPolicy.getImagePoint());
                 rewardPointForContent = rewardPointForContent.add(reviewPointPolicy.getContentPoint());
             }
         }
         if(isReviewImage) {
-            reviewPoint = reviewPoint.add(rewardPointForImage).add(rewardPointForContent);
+            reviewPoint = reviewPoint.add(rewardPointForImage).add(rewardPointForContent).multiply(BigDecimal.valueOf(count));
         }
         else {
-            reviewPoint = reviewPoint.add(rewardPointForContent);
+            reviewPoint = reviewPoint.add(rewardPointForContent).multiply(BigDecimal.valueOf(count));
         }
 
         return reviewPoint;
     }
 
+    // 리뷰 포인트 정책 전제 조회
     public List<PointPolicyResDto.ReviewPointPolicy> findReviewPointRange() {
-        List<ReviewPointPolicy> reviewPointPolicyList = new ArrayList<>();
-        reviewPointPolicyList.add(ReviewPointPolicy.REVIEW_RANGE_1);
-        reviewPointPolicyList.add(ReviewPointPolicy.REVIEW_RANGE_2);
-        reviewPointPolicyList.add(ReviewPointPolicy.REVIEW_RANGE_3);
-        reviewPointPolicyList.add(ReviewPointPolicy.REVIEW_RANGE_4);
-
+        List<ReviewPointPolicy> reviewPointPolicyList = List.of(ReviewPointPolicy.class.getEnumConstants());
         return reviewPointPolicyList.stream().map(pointMapper::toReviewPointPolicyResponseDto).collect(Collectors.toList());
     }
 
+    // 이벤트 삭제
+    @Transactional
     public void deletePointHistoryByPointPolicy(PointPolicy pointPolicy) {
         // point 가 0 인 로그만 가져와서 다 지움
         List<PointHistory> pointHistoryList = qPointHistoryRepository.findAllByPointPolicy(pointPolicy);
+        if(pointHistoryList.isEmpty()) return;
         pointHistoryRepository.deleteAll(pointHistoryList);
     }
 
-    public void createPointHistoryByPointPolicy(User user, PointPolicy pointPolicy, Map<String, BigInteger> ids) {
+    @Transactional
+    public void createPointHistoryByPointPolicy(User user, PointPolicy pointPolicy, BigInteger noticeId, PointStatus pointStatus) {
+        // 정책에 해당하는 조건 가져오기
         Integer completeCount = pointPolicy.getCompletedConditionCount();
         Integer accountLimit = pointPolicy.getAccountCompletionLimit();
 
+        // 정책 id를 가졌고, 유저가 동일한 포인트 내역 조회
         List<PointHistory> pointHistoryList = qPointHistoryRepository.findAllByUserAndPointPolicy(user, pointPolicy);
+        // 이벤트 참여가 처음이면
         if(pointHistoryList.isEmpty()) {
-            PointHistory pointHistory = pointMapper.createPointHistoryForCount(user, pointPolicy, ids, 0);
+            PointHistory pointHistory = pointMapper.createPointHistoryForCount(user, pointPolicy, noticeId, 0, pointStatus);
             pointHistoryRepository.save(pointHistory);
             return;
         }
 
+        // 이벤트 참여 횟수
         List<PointHistory> rewardPointHistoryCount = pointHistoryList.stream().filter(pointHistory -> pointHistory.getPoint().equals(BigDecimal.valueOf(0))).toList();
-        PointHistory pointHistory = pointMapper.createPointHistoryForCount(user, pointPolicy, ids, rewardPointHistoryCount.size());
+        PointHistory pointHistory = pointMapper.createPointHistoryForCount(user, pointPolicy, noticeId, rewardPointHistoryCount.size(), pointStatus);
 
         pointHistoryRepository.save(pointHistory);
     }
 
-    public void createPointHistoryByReview(User user, BigInteger id, BigDecimal point) {
-        pointHistoryRepository.save(pointMapper.createPointHistoryForReview(user, id, point));
+    @Transactional
+    public void createPointHistoryByOthers(User user, BigInteger id, PointStatus pointStatus, BigDecimal point) {
+        PointHistory pointHistory = pointMapper.createPointHistoryByOthers(user, id, pointStatus, point);
+        pointHistoryRepository.save(pointHistory);
     }
 }
